Big changes in structure
Removed the table output of function loadDrinks() in favour of presenting information on page with images. Played with the .hide() and .show() functionality of buttons. Changed the JSON loading to happen upon page load. There are now proper images and links to recipes on our website, they just need filling out.